fix: Address additional GitHub Copilot review feedback for PR #331
1. Moved test utilities into test_serve_e2e.py
   - Removed brittle sys.path manipulation
   - Helper functions now in the file that uses them
   - Keeps conftest.py focused on fixtures only

2. Tightened stop server verification
   - Now requires 200 response with state=stopped
   - Prevents masking regressions with 404 acceptance

3. Fixed tools endpoint response parsing
   - Handles both dict and list response formats
   - Compatible with current API contract

4. Removed unused variables and imports
   - Prefixed unused fixture unpacking with underscore
   - Removed old_pid variable and unused imports

All changes validated with passing tests.
