+++ tests/conftest.py
@@ -116,74 +116,3 @@ async def api_client(serve_test_app):
     transport = httpx.ASGITransport(app=app)
     async with httpx.AsyncClient(transport=transport, base_url="http://test") as client:
         yield client
-
-
-# ============================================================================
-# Test Utility Functions
-# ============================================================================
-
-async def wait_for_server_status(client, server_id: str, expected_status: str, timeout: float = 10.0) -> bool:
-    """
-    Poll server status until it reaches expected state or timeout.
-
-    This utility replaces hardcoded asyncio.sleep() calls with intelligent polling,
-    making tests more reliable and faster.
-
-    Args:
-        client: httpx.AsyncClient instance
-        server_id: The server ID to check
-        expected_status: The status to wait for (e.g., "running", "stopped")
-        timeout: Maximum time to wait in seconds
-
-    Returns:
-        True if server reached expected status, False if timeout
-
-    Example:
-        success = await wait_for_server_status(client, "test-server", "running", timeout=30)
-        assert success, "Server failed to start within 30 seconds"
-    """
-    start = time.time()
-    while time.time() - start < timeout:
-        try:
-            response = await client.get(f"/api/servers/{server_id}/status")
-            if response.status_code == 200:
-                data = response.json()
-                # Handle both 'state' and 'status' field names (API inconsistency)
-                status = data.get("state") or data.get("status")
-                if status == expected_status:
-                    return True
-        except Exception:
-            pass  # Continue polling on transient errors
-        await asyncio.sleep(0.5)  # Poll every 500ms
-    return False
-
-
-async def wait_for_condition(condition_func, timeout: float = 10.0, poll_interval: float = 0.5) -> bool:
-    """
-    Generic polling utility for waiting on async conditions.
-
-    Args:
-        condition_func: Async function that returns True when condition is met
-        timeout: Maximum time to wait in seconds
-        poll_interval: Time between checks in seconds
-
-    Returns:
-        True if condition met, False if timeout
-
-    Example:
-        async def check_health():
-            response = await client.get("/health")
-            return response.status_code == 200
-
-        success = await wait_for_condition(check_health, timeout=30)
-        assert success, "Health check never succeeded"
-    """
-    start = time.time()
-    while time.time() - start < timeout:
-        try:
-            if await condition_func():
-                return True
-        except Exception:
-            pass  # Continue polling on errors
-        await asyncio.sleep(poll_interval)
-    return False
